+++ ntex/src/server/config.rs
@@ -1,6 +1,6 @@
 use std::{
-    cell::RefCell, fmt, future::Future, io, marker::PhantomData, mem, net, pin::Pin,
-    rc::Rc,
+    cell::Cell, cell::RefCell, fmt, future::Future, io, marker::PhantomData, mem, net,
+    pin::Pin, rc::Rc,
 };
 
 use log::error;
@@ -14,7 +14,32 @@ use super::service::{
 };
 use super::Token;
 
-pub struct ServiceConfig {
+#[derive(Clone)]
+pub struct ServiceConfig(pub(super) Rc<InnerServiceConfig>);
+
+pub(super) struct InnerServiceConfig {
+    pub(super) pool: Cell<PoolId>,
+}
+
+impl Default for ServiceConfig {
+    fn default() -> Self {
+        Self(Rc::new(InnerServiceConfig {
+            pool: Cell::new(PoolId::DEFAULT),
+        }))
+    }
+}
+
+impl ServiceConfig {
+    /// Set memory pool for the service.
+    ///
+    /// Use specified memory pool for memory allocations.
+    pub fn memory_pool(&self, id: PoolId) -> &Self {
+        self.0.pool.set(id);
+        self
+    }
+}
+
+pub struct Configuration {
     pub(super) services: Vec<(String, net::TcpListener)>,
     pub(super) apply: Box<dyn ServiceRuntimeConfiguration + Send>,
     pub(super) threads: usize,
@@ -22,9 +47,9 @@ pub struct ServiceConfig {
     applied: bool,
 }
 
-impl ServiceConfig {
-    pub(super) fn new(threads: usize, backlog: i32) -> ServiceConfig {
-        ServiceConfig {
+impl Configuration {
+    pub(super) fn new(threads: usize, backlog: i32) -> Self {
+        Configuration {
             threads,
             backlog,
             services: Vec::new(),
@@ -78,7 +103,7 @@ impl ServiceConfig {
     /// It get executed in the worker thread.
     pub fn on_worker_start<F, R, E>(&mut self, f: F) -> io::Result<()>
     where
-        F: Fn(ServiceRuntime) -> R + Send + Clone + 'static,
+        F: Fn(RuntimeConfiguration) -> R + Send + Clone + 'static,
         R: Future<Output = Result<(), E>> + 'static,
         E: fmt::Display + 'static,
     {
@@ -126,15 +151,13 @@ impl InternalServiceFactory for ConfiguredService {
         })
     }
 
-    fn set_memory_pool(&self, _: &str, _: PoolId) {}
-
     fn create(
         &self,
     ) -> Pin<Box<dyn Future<Output = Result<Vec<(Token, BoxedServerService)>, ()>>>>
     {
         // configure services
-        let rt = ServiceRuntime::new(self.topics.clone());
-        let cfg_fut = self.rt.configure(ServiceRuntime(rt.0.clone()));
+        let rt = RuntimeConfiguration::new(self.topics.clone());
+        let cfg_fut = self.rt.configure(RuntimeConfiguration(rt.0.clone()));
         let mut names = self.names.clone();
         let tokens = self.services.clone();
 
@@ -185,7 +208,7 @@ pub(super) trait ServiceRuntimeConfiguration {
 
     fn configure(
         &self,
-        rt: ServiceRuntime,
+        rt: RuntimeConfiguration,
     ) -> Pin<Box<dyn Future<Output = Result<(), ()>>>>;
 }
 
@@ -199,7 +222,7 @@ unsafe impl<F: Send, R, E> Send for ConfigWrapper<F, R, E> {}
 
 impl<F, R, E> ServiceRuntimeConfiguration for ConfigWrapper<F, R, E>
 where
-    F: Fn(ServiceRuntime) -> R + Send + Clone + 'static,
+    F: Fn(RuntimeConfiguration) -> R + Send + Clone + 'static,
     R: Future<Output = Result<(), E>> + 'static,
     E: fmt::Display + 'static,
 {
@@ -212,7 +235,7 @@ where
 
     fn configure(
         &self,
-        rt: ServiceRuntime,
+        rt: RuntimeConfiguration,
     ) -> Pin<Box<dyn Future<Output = Result<(), ()>>>> {
         let f = self.f.clone();
         Box::pin(async move {
@@ -227,7 +250,7 @@ fn not_configured() {
     error!("Service is not configured");
 }
 
-pub struct ServiceRuntime(Rc<RefCell<ServiceRuntimeInner>>);
+pub struct RuntimeConfiguration(Rc<RefCell<ServiceRuntimeInner>>);
 
 struct ServiceRuntimeInner {
     names: HashMap<String, Token>,
@@ -235,9 +258,9 @@ struct ServiceRuntimeInner {
     onstart: Vec<Pin<Box<dyn Future<Output = ()>>>>,
 }
 
-impl ServiceRuntime {
+impl RuntimeConfiguration {
     fn new(names: HashMap<String, Token>) -> Self {
-        ServiceRuntime(Rc::new(RefCell::new(ServiceRuntimeInner {
+        RuntimeConfiguration(Rc::new(RefCell::new(ServiceRuntimeInner {
             names,
             services: HashMap::default(),
             onstart: Vec::new(),