+++ ntex/src/server/mod.rs
@@ -14,8 +14,7 @@ mod worker;
 
 pub(crate) use self::builder::create_tcp_listener;
 pub use self::builder::ServerBuilder;
-pub use self::config::{ServiceConfig, ServiceRuntime};
-pub use self::service::StreamServiceFactory;
+pub use self::config::{Configuration, RuntimeConfiguration, ServiceConfig};
 pub use self::test::{build_test_server, test_server, TestServer};
 
 #[non_exhaustive]